refactor: enforce max-lines: 300 lint, split oversized files
Adds oxlint max-lines rule (300, skip blank/comments) with overrides
for test files and packages/tool-registry/src/generated/**, and
splits the three source files that exceeded the threshold:

- tools/image-resizer/client.tsx (683 → 89): extracted types,
  constants, useImageResizer hook, and Upload/Options/Result cards
- packages/tool-registry/src/generate.ts (471 → 76): split into
  generate/{io,load,codegen}.ts
- tools/json-schema-validator/client.tsx (432 → 258): extracted
  types, constants, and ResultCard
