+++ tools/json-schema-validator/client.tsx
@@ -7,11 +7,7 @@ import {
 } from "react"
 
 import { ToolCopyButton } from "@workspace/ui/components/tool/tool-copy-button"
-import {
-  Alert,
-  AlertDescription,
-  AlertTitle,
-} from "@workspace/ui/components/ui/alert"
+import { Badge } from "@workspace/ui/components/ui/badge"
 import { Button } from "@workspace/ui/components/ui/button"
 import {
   Card,
@@ -23,95 +19,20 @@ import {
 } from "@workspace/ui/components/ui/card"
 import { Label } from "@workspace/ui/components/ui/label"
 import { Switch } from "@workspace/ui/components/ui/switch"
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@workspace/ui/components/ui/table"
 import { Textarea } from "@workspace/ui/components/ui/textarea"
-import { Badge } from "@workspace/ui/components/ui/badge"
-import {
-  BadgeCheck,
-  FileJson2,
-  RefreshCcw,
-  TriangleAlert,
-} from "@workspace/ui/icons"
+import { FileJson2, RefreshCcw } from "@workspace/ui/icons"
 
+import { DEFAULT_DATA, DEFAULT_SCHEMA, STORAGE_KEYS } from "./client/constants"
+import { ResultCard } from "./client/result-card"
+import type { JsonSchemaMessages } from "./client/types"
 import {
   validateJsonSchemaText,
   type ValidationResult,
 } from "./core/validate-json-schema"
 
-type JsonSchemaMessages = Readonly<{
-  meta: {
-    name: string
-    description: string
-  }
-  schemaLabel: string
-  schemaDescription: string
-  schemaPlaceholder: string
-  dataLabel: string
-  dataDescription: string
-  dataPlaceholder: string
-  optionsTitle: string
-  optionsDescription: string
-  validateFormatsLabel: string
-  validateFormatsDescription: string
-  allErrorsLabel: string
-  allErrorsDescription: string
-  draftLabel: string
-  idleTitle: string
-  idleDescription: string
-  parseErrorTitle: string
-  schemaErrorTitle: string
-  validTitle: string
-  validDescription: string
-  invalidTitle: string
-  invalidDescription: string
-  errorPathLabel: string
-  errorKeywordLabel: string
-  errorMessageLabel: string
-  resultTitle: string
-  resultDescription: string
-  copySchemaLabel: string
-  copyDataLabel: string
-  copyErrorsLabel: string
-  copiedLabel: string
-  loadExampleLabel: string
-  clearLabel: string
-}>
-
 type JsonSchemaValidatorClientProps = Readonly<{
   messages: JsonSchemaMessages
 }>
-
-const DEFAULT_SCHEMA = `{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "type": "object",
-  "properties": {
-    "id": { "type": "string", "format": "uuid" },
-    "name": { "type": "string", "minLength": 1 },
-    "age": { "type": "integer", "minimum": 0 }
-  },
-  "required": ["id", "name"],
-  "additionalProperties": false
-}`
-
-const DEFAULT_DATA = `{
-  "id": "1b9d6bcd-bbfd-4b2d-9b5d-ab8dfbbd4bed",
-  "name": "Ada Lovelace",
-  "age": 37
-}`
-
-const STORAGE_KEYS = {
-  allErrors: "tools:json-schema-validator:all-errors",
-  data: "tools:json-schema-validator:data",
-  schema: "tools:json-schema-validator:schema",
-  validateFormats: "tools:json-schema-validator:validate-formats",
-} as const
 
 function JsonSchemaValidatorClient({
   messages,
@@ -330,101 +251,11 @@ function JsonSchemaValidatorClient({
         </Card>
       </div>
 
-      <Card>
-        <CardHeader>
-          <CardTitle>{messages.resultTitle}</CardTitle>
-          <CardDescription>{messages.resultDescription}</CardDescription>
-        </CardHeader>
-        <CardContent className="flex flex-col gap-4">
-          {validation.state === "idle" ? (
-            <Alert>
-              <FileJson2 />
-              <AlertTitle>{messages.idleTitle}</AlertTitle>
-              <AlertDescription>{messages.idleDescription}</AlertDescription>
-            </Alert>
-          ) : null}
-
-          {validation.state === "parse-error" ? (
-            <Alert variant="destructive">
-              <TriangleAlert />
-              <AlertTitle>{messages.parseErrorTitle}</AlertTitle>
-              <AlertDescription>
-                <span className="font-medium">
-                  {validation.source === "schema"
-                    ? messages.schemaLabel
-                    : messages.dataLabel}
-                  :
-                </span>{" "}
-                {validation.message}
-              </AlertDescription>
-            </Alert>
-          ) : null}
-
-          {validation.state === "schema-error" ? (
-            <Alert variant="destructive">
-              <TriangleAlert />
-              <AlertTitle>{messages.schemaErrorTitle}</AlertTitle>
-              <AlertDescription>{validation.message}</AlertDescription>
-            </Alert>
-          ) : null}
-
-          {validation.state === "validated" && validation.valid ? (
-            <Alert>
-              <BadgeCheck />
-              <AlertTitle>{messages.validTitle}</AlertTitle>
-              <AlertDescription>{messages.validDescription}</AlertDescription>
-            </Alert>
-          ) : null}
-
-          {validation.state === "validated" && !validation.valid ? (
-            <>
-              <Alert variant="destructive">
-                <TriangleAlert />
-                <AlertTitle>{messages.invalidTitle}</AlertTitle>
-                <AlertDescription>
-                  {messages.invalidDescription}
-                </AlertDescription>
-              </Alert>
-
-              <div className="rounded-xl border border-border/70">
-                <Table>
-                  <TableHeader>
-                    <TableRow>
-                      <TableHead>{messages.errorPathLabel}</TableHead>
-                      <TableHead>{messages.errorKeywordLabel}</TableHead>
-                      <TableHead>{messages.errorMessageLabel}</TableHead>
-                    </TableRow>
-                  </TableHeader>
-                  <TableBody>
-                    {validation.issues.map((issue) => (
-                      <TableRow
-                        key={`${issue.path}:${issue.keyword}:${issue.message}`}
-                      >
-                        <TableCell className="font-mono text-xs text-muted-foreground">
-                          {issue.path}
-                        </TableCell>
-                        <TableCell className="font-mono text-xs text-muted-foreground">
-                          {issue.keyword}
-                        </TableCell>
-                        <TableCell>{issue.message}</TableCell>
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              </div>
-            </>
-          ) : null}
-        </CardContent>
-        {errorsJson ? (
-          <CardFooter className="justify-end">
-            <ToolCopyButton
-              value={errorsJson}
-              copyLabel={messages.copyErrorsLabel}
-              copiedLabel={messages.copiedLabel}
-            />
-          </CardFooter>
-        ) : null}
-      </Card>
+      <ResultCard
+        errorsJson={errorsJson}
+        messages={messages}
+        validation={validation}
+      />
     </div>
   )
 }
