+++ tools/json-schema-validator/client/types.ts
@@ -1,0 +1,38 @@
+export type JsonSchemaMessages = Readonly<{
+  meta: {
+    name: string
+    description: string
+  }
+  schemaLabel: string
+  schemaDescription: string
+  schemaPlaceholder: string
+  dataLabel: string
+  dataDescription: string
+  dataPlaceholder: string
+  optionsTitle: string
+  optionsDescription: string
+  validateFormatsLabel: string
+  validateFormatsDescription: string
+  allErrorsLabel: string
+  allErrorsDescription: string
+  draftLabel: string
+  idleTitle: string
+  idleDescription: string
+  parseErrorTitle: string
+  schemaErrorTitle: string
+  validTitle: string
+  validDescription: string
+  invalidTitle: string
+  invalidDescription: string
+  errorPathLabel: string
+  errorKeywordLabel: string
+  errorMessageLabel: string
+  resultTitle: string
+  resultDescription: string
+  copySchemaLabel: string
+  copyDataLabel: string
+  copyErrorsLabel: string
+  copiedLabel: string
+  loadExampleLabel: string
+  clearLabel: string
+}>
